move baseHandler to dedicated file
